Restructure files from template
