+++ ton-drive-frontend/src/features/demo/ui/Jetton.tsx
@@ -1,5 +1,5 @@
 import { beginCell, toNano, Address, Cell, fromNano } from "ton";
-import { useTonConnect } from "../hooks/useTonConnect";
+import { useTonConnect } from "../../../shared/hooks/useTonConnect";
 import { useFaucetJettonContract } from "../hooks/useFaucetJettonContract";
 import {
   Card,
@@ -7,7 +7,7 @@ import {
   FlexBoxRow,
   Button,
   Ellipsis,
-} from "./styled/styled";
+} from "../../../shared/ui/styled";
 
 export function Jetton() {
   const { connected } = useTonConnect();
