+++ ton-drive-frontend/src/shared/hooks/useAsyncInitialize.ts
