+++ ton-drive-frontend/src/shared/ui/styled.tsx
